Use maven-dependency-plugin for finding artifacts
This change introduces the maven dependency plugin in order to resolve project dependencies.
This seems to work despite the changes in the enforcer plugin in 1.3.1. It is an
experimental fix for issue #13. However there are a number of scenarios that still need
to be examined. I am concerned that this fix will break backwards compatibility with older
maven versions. So more testing is needed before I will be happy to merge this into the
master branch ( and do a release ).

The sample project has be updated to include this snapshot as a dependency (as well as
updating the enforcer plugin version to 1.3.1. As I have not released a snapshot as yet
you will need to run mvn clean install on the victims-enforcer project to test these
changes against the sample project.

A couple of other minor changes. The tests were polluting the users victims database. This
is not acceptable. The database has been moved to a temporary directory on the
file system and the http webservice has been mocked out (already was). I've also
introduced the convention of prepending debug output with [victims-enforcer]. This
makes it easier to grep for debug messages.  I will have to refactor the logging
to provide a more sane approach to doing this at some stage.

- Grant.